# Conflicts:
#	WalkieTalkie/R.generated.swift

+++ WalkieTalkie/WalkieTalkie/UI/AmongUs/Home/Dm/AmongChat.Home.ConversationListController.swift
@@ -22,6 +22,10 @@ extension AmongChat.Home {
                 .flatMap { item -> Single<[Entity.DMConversation]> in
                     return DMManager.shared.conversations()
                 }
+                .do(onNext: { items in
+                    let unreadCount = items.reduce(0, { $0 + $1.unreadCount })
+                    Settings.shared.hasUnreadMessageRelay.accept(unreadCount > 0)
+                })
                 .observeOn(MainScheduler.asyncInstance)
                 .bind(to: dataSourceReplay)
                 .disposed(by: bag)
@@ -37,25 +41,6 @@ extension AmongChat.Home {
         }
         
         private lazy var navigationView = AmongChat.Home.NavigationBar(.notice)
-        
-//        private lazy var collectionView: UICollectionView = {
-//            let layout = UICollectionViewFlowLayout()
-//            layout.scrollDirection = .vertical
-//            var hInset: CGFloat = 20
-//            var columns: Int = 1
-//            let interitemSpacing: CGFloat = 20
-//            layout.itemSize = CGSize(width: UIScreen.main.bounds.width, height: 84)
-//            layout.sectionInset = UIEdgeInsets(top: 12, left: hInset, bottom: 0, right: hInset)
-//            let v = UICollectionView(frame: .zero, collectionViewLayout: layout)
-//            v.register(nibWithCellClass: ConversationListCell.self)
-//            v.showsVerticalScrollIndicator = false
-//            v.showsHorizontalScrollIndicator = false
-//            v.dataSource = self
-//            v.delegate = self
-//            v.backgroundColor = .clear
-//            v.alwaysBounceVertical = true
-//            return v
-//        }()
         
         private lazy var listView: UITableView = {
             let v = UITableView(frame: .zero, style: .plain)
@@ -78,18 +63,7 @@ extension AmongChat.Home {
         }()
         
         private let hasUnreadNotice = BehaviorRelay(value: false)
-                
-//        var dataSource: [Entity.Notice] = [] {
-//            didSet {
-////                noticeVMList = dataSource.enumerated().map({ [weak self] (idx, notice) in
-////                    NoticeViewModel(with: notice) {
-////                        self?.noticeListView.reloadItems(at: [IndexPath(item: idx, section: 0)])
-////                    }
-////
-////                })
-//            }
-//        }
-        
+                        
         private var dataSource: [Entity.DMConversation] = [] {
             didSet {
                 listView.reloadData()
@@ -120,12 +94,6 @@ extension AmongChat.Home.ConversationListController: UITableViewDataSource, UITa
         let cell = tableView.dequeueReusableCell(withClass: ConversationTableCell.self, for: indexPath)
         if let item = dataSource.safe(indexPath.row) {
             cell.bind(item)
-//            cell.configView(with: user, isFollowing: false, isSelf: false)
-//            cell.updateFollowData = { [weak self] (follow) in
-//                guard let `self` = self else { return }
-//                self.userList[indexPath.row].isFollowed = follow
-//                self.addLogForFollow(with: self.userList[indexPath.row].uid)
-//            }
         }
         return cell
     }
@@ -139,7 +107,7 @@ extension AmongChat.Home.ConversationListController: UITableViewDataSource, UITa
     }
     
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return 0.001
+        return 6
     }
     
     func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
@@ -164,8 +132,15 @@ extension AmongChat.Home.ConversationListController: UITableViewDataSource, UITa
             return nil
         }
         let action = UIContextualAction(style: .destructive, title: R.string.localizable.amongChatDelete()) { [weak self] action, view, handler in
-            handler(true)
-            self?.deleteAllHistory(for: item.fromUid)
+            guard let `self` = self else { return }
+            self.showAmongAlert(title: R.string.localizable.dmDeleteHistoryAlertTitle(),
+                                cancelTitle: R.string.localizable.toastCancel(),
+                                confirmTitle: R.string.localizable.groupRoomYes(),
+                                confirmTitleColor: "#FB5858".color(),
+                                confirmAction: { [weak self] in
+                                    handler(true)
+                                    self?.deleteAllHistory(for: item.fromUid)
+                                })
         }
         action.image = R.image.iconDmConversationDelete()
         action.backgroundColor = .red
@@ -182,57 +157,11 @@ extension AmongChat.Home.ConversationListController: UITableViewDataSource, UITa
     }
 }
 
-extension AmongChat.Home.ConversationListController: UICollectionViewDataSource {
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return dataSource.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let item = dataSource.safe(indexPath.item) else {
-            return UICollectionViewCell()
-        }
-        
-        let cell = collectionView.dequeueReusableCell(withClass: ConversationListCell.self, for: indexPath)
-        
-        cell.bind(item)
-//        switch notice.notice.message.messageType {
-//        case .TxtMsg, .ImgMsg, .ImgTxtMsg, .TxtImgMsg:
-//            cell = collectionView.dequeueReusableCell(withClass: ConversationListCell.self, for: indexPath)
-//            if let cell = cell as? ConversationListCell {
-////                cell.bindNoticeData(notice)
-//            }
-
-//        case .SocialMsg:
-//            cell = collectionView.dequeueReusableCell(withReuseIdentifier: NSStringFromClass(SocialMessageCell.self), for: indexPath)
-//
-//            if let cell = cell as? SocialMessageCell {
-//                cell.bindNoticeData(notice)
-//            }
-//
-//        }
-        
-        return cell
-    }
-}
-
-extension AmongChat.Home.ConversationListController: UICollectionViewDelegate {
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        guard let item = dataSource.safe(indexPath.item) else {
-            return
-        }
-        let vc = ConversationViewController(item)
-        navigationController?.pushViewController(vc)
-    }
-    
-}
-
 extension AmongChat.Home.ConversationListController {
     
     func deleteAllHistory(for uid: String) {
         let removeBlock = view.raft.show(.loading)
-        DMManager.shared.clearAllMessage(of: uid)
+        DMManager.shared.deleteConversation(of: uid)
             .subscribe(onSuccess: { [weak self] in
                 removeBlock()
 //                self?.navigationController?.popViewController()
